+++ src/app/core/http/api/api.service.ts
@@ -5,6 +5,7 @@ import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import { catchError } from 'rxjs/operators';
 
 import { Institution } from 'src/app/core/models/institution';
+import { APIErrorResponse } from 'src/app/core/models/error';
 import { Offer } from 'src/app/core/models/offer';
 import { User } from 'src/app/core/models/user';
 import {
@@ -16,11 +17,7 @@ import {
   OfferSearchFilterToAPI,
 } from './api.interfaces';
 import { AuthResponseData } from 'src/app/core/auth/auth.interfaces';
-import { LogService } from 'src/app/core/services/logger/log.service';
-import {
-  ErrorCodes,
-  ErrorHandlerService,
-} from 'src/app/core/services/error-handling/error-handling';
+import { ErrorCodes } from 'src/app/core/services/error-handling/error-handling';
 
 /**
  * api.service.ts
@@ -30,24 +27,25 @@ import {
  * Latest update 17.02.2022/ml
  */
 
-/* eslint-disable no-console */
+
+export const TOKEN_PATH = '/oauth/token';
+export const LOGOUT_PATH = '/api/logout';
+export const MAX_TOKEN_REFRESH = 3;
 
 @Injectable({
   providedIn: 'root',
 })
 export class ApiService {
-  constructor(
-    private http: HttpClient,
-    private logService: LogService,
-    private errorHandler: ErrorHandlerService
-  ) {}
+  private countTokenCall = 0;
+  constructor(private http: HttpClient) {}
 
   ////////////////////////////////////////////////
   // Authenticate
   ////////////////////////////////////////////////
   public loginUser(email: string, password: string): Observable<AuthResponseData> {
+    this.countTokenCall = 0;
     return this.http
-      .post<AuthResponseData>(environment.apiURL + '/oauth/token', {
+      .post<AuthResponseData>(environment.apiURL + TOKEN_PATH, {
         grant_type: environment.clientLoginData.grantType,
         username: email,
         password: password,
@@ -66,7 +64,7 @@ export class ApiService {
    * @returns Observable<any>
    */
   public logoutUser(): Observable<any> {
-    return this.http.get<any>(environment.apiURL + '/api/logout').pipe(
+    return this.http.get<any>(environment.apiURL + LOGOUT_PATH).pipe(
       catchError((errorResponse: HttpErrorResponse) => {
         return this.handleError(errorResponse);
       })
@@ -79,18 +77,25 @@ export class ApiService {
    * @returns
    */
   public updateUserSession(refreshToken: string): Observable<any> {
-    return this.http
-      .post<AuthResponseData>(environment.apiURL + '/oauth/token', {
+    // Wichtig, hier kein Fehlerhandling, das übernimmt der Interceptor
+    // Die Abfrage nach der Anzahl der Request sperrt mehr als x Abfragen
+    if (++this.countTokenCall > MAX_TOKEN_REFRESH) {
+      return throwError(
+        () =>
+          new APIErrorResponse({
+            error: { error: 'refresh_failure' },
+            status: 429,
+            message: 'refresh_failure',
+          })
+      );
+    } else {
+      return this.http.post<AuthResponseData>(environment.apiURL + TOKEN_PATH, {
         grant_type: 'refresh_token',
         refresh_token: refreshToken,
         client_id: environment.clientLoginData.clientId,
         client_secret: environment.clientLoginData.clientSecret,
-      })
-      .pipe(
-        catchError((errorResponse: HttpErrorResponse) => {
-          return this.handleError(errorResponse);
-        })
-      );
+      });
+    }
   }
 
   ////////////////////////////////////////////////
@@ -403,7 +408,7 @@ export class ApiService {
   // ErrorHandling
   ////////////////////////////////////////////////
   // Todo: Fehlermeldungstexte für die GUI
-  private handleError(errorRes: HttpErrorResponse): Observable<never> {
+  private handleError(errorRes: HttpErrorResponse | APIErrorResponse): Observable<never> {
     let errorCode = ErrorCodes.UNKNOWN;
 
     switch (errorRes.status) {
@@ -414,7 +419,11 @@ export class ApiService {
         errorCode = ErrorCodes.E400;
         break;
       case 401:
-        errorCode = ErrorCodes.E401;
+        if (errorRes.error && errorRes.error.error === 'invalid_request') {
+          errorCode = ErrorCodes.E401_1;
+        } else {
+          errorCode = ErrorCodes.E401;
+        }
         break;
       case 403:
         errorCode = ErrorCodes.E403;
@@ -426,7 +435,11 @@ export class ApiService {
         errorCode = ErrorCodes.E422;
         break;
       case 429:
-        errorCode = ErrorCodes.E429;
+        if (errorRes.error && errorRes.message === 'refresh_failure') {
+          errorCode = ErrorCodes.E429_1;
+        } else {
+          errorCode = ErrorCodes.E429;
+        }
         break;
       case 500:
         errorCode = ErrorCodes.E500;
@@ -434,7 +447,7 @@ export class ApiService {
     }
     let newError = new Error(errorCode);
 
-    console.log('ErrorResponse:', errorRes);
+    //console.log('ErrorResponse:', errorRes);
     return throwError(() => newError);
   }
 }